Separate test modules from their corresponding modules.
This makes a lot more files, but makes the library code a lot easier to read (since there's significantly less of it).
